- Setup test with jest
