+++ auth/src/app.ts
@@ -1,0 +1,36 @@
+import express from 'express';
+import 'express-async-errors';
+import cookieSession from 'cookie-session';
+
+import { currentUserRouter } from './routes/current-user';
+import { signinRouter } from './routes/signin';
+import { signoutRouter } from './routes/signout';
+import { signupRouter } from './routes/signup';
+import { errorHandler } from './middlewares/error-handler';
+import { NotFoundError } from './errors/notfound-error';
+
+const app = express();
+app.set('trust proxy', true);
+app.use(express.json());
+app.use(
+  cookieSession({
+    signed: false,
+    secure: true,
+  })
+);
+
+app.get('/api/users/healthcheck', (req, res) => {
+  res.send('Auth service is up and running....');
+});
+app.use(currentUserRouter);
+app.use(signinRouter);
+app.use(signoutRouter);
+app.use(signupRouter);
+app.all('*', async () => {
+  throw new NotFoundError();
+});
+
+app.use(errorHandler);
+
+
+export { app };
