+++ auth/src/routes/__test__/signup.test.ts
@@ -1,0 +1,12 @@
+import request from 'supertest';
+import { app } from '../../app';
+
+it('returns 201 on successful signup', async () => {
+  return request(app)
+    .post('/api/users/signup')
+    .send({
+      email: 'test@test.com',
+      password: 'password',
+    })
+    .expect(201);
+});
